# secureprompt-schemas/openapi/v1/openapi.yaml
openapi: 3.0.3
info:
  title: SecurePrompt Gateway API
  version: "1.0.0"
  description: >
    SecurePrompt is an LLM security gateway that sits between enterprise
    applications and AI providers. It redacts PII/secrets, enforces policy,
    tracks token costs, and provides governance analytics.

    Two auth schemes are used:
    - **JWT** (`Authorization: Bearer eyJ...`) — dashboard routes
    - **API Key** (`Authorization: Bearer sp-...`) — gateway + MCP utility routes
servers:
  - url: http://localhost:8080
    description: Local dev Rust gateway
  - url: https://api.secureprompt.tech
    description: Cloud
...
paths:
...
  /v1/providers:
    get:
      tags: [Providers]
      summary: List LLM providers
      operationId: listProviders
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of providers (credential never included)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProviderResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Providers]
      summary: Create a provider (admin only)
      operationId: createProvider
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProviderRequest"
      responses:
        "201":
          description: Provider created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/providers/{id}:
    put:
      tags: [Providers]
      summary: Update / rotate provider credential (admin only)
      operationId: updateProvider
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProviderRequest"
      responses:
        "200":
          description: Provider updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Providers]
      summary: Delete a provider (admin only)
      operationId: deleteProvider
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "204":
          description: Provider deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Policy Rules ──────────────────────────────────────────────────────────────
...
  /v1/secure-mode:
    get:
      tags: [Settings]
      summary: Get secure mode configuration
      operationId: getSecureMode
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current secure mode configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecureMode"
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
      tags: [Settings]
      summary: Update secure mode configuration
      operationId: putSecureMode
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SecureModeRequest"
      responses:
        "200":
          description: Secure mode configuration updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecureMode"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Dashboard JWT — Authorization: Bearer eyJ..."
    apiKeyAuth:
      type: http
      scheme: bearer
      description: "API key — Authorization: Bearer sp-..."
  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Forbidden:
      description: Insufficient role or workspace mismatch
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
  schemas:
...
    ProviderResponse:
      type: object
      required: [id, name, provider_type, has_credential, last_rotated_at, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        provider_type:
          type: string
          example: openai
        has_credential:
          type: boolean
          description: True when an encrypted credential is stored
        last_rotated_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    CreateProviderRequest:
      type: object
      required: [name, provider_type]
      properties:
        name:
          type: string
        provider_type:
          type: string
          example: openai
        credential:
          type: string
          nullable: true
          description: Plaintext API key — encrypted before storage
    UpdateProviderRequest:
      type: object
      properties:
        name:
          type: string
          nullable: true
        provider_type:
          type: string
          nullable: true
        credential:
          type: string
          nullable: true
          description: New plaintext key; omit to leave existing unchanged

    # ── Policy Rules ─────────────────────────────────────────────────────────────
...
    SecureMode:
      type: object
      required:
        - enabled
        - level
        - block_on_pii_detection
        - block_on_injection_detection
        - redact_pii_in_responses
        - updated_at
      properties:
        enabled:
          type: boolean
        level:
          type: string
          enum: [permissive, standard, strict]
        block_on_pii_detection:
          type: boolean
        block_on_injection_detection:
          type: boolean
        redact_pii_in_responses:
          type: boolean
        updated_at:
          type: string
          format: date-time
    SecureModeRequest:
      type: object
      properties:
        enabled:
          type: boolean
        level:
          type: string
          enum: [permissive, standard, strict]
        block_on_pii_detection:
          type: boolean
        block_on_injection_detection:
          type: boolean
        redact_pii_in_responses:
          type: boolean
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message, type]
          properties:
            message:
              type: string
            type:
              type: string
              example: secureprompt_error
            code:
              type: string
              description: Stable machine-readable code (e.g. budget_exceeded)
...